Using proper 64bit types here

+++ src/engine/i_audio.c
@@ -1106,10 +1106,10 @@ static void Seq_Shutdown(doomseq_t* seq) {
 
 static int SDLCALL Thread_PlayerHandler(void* param) {
     doomseq_t* seq = (doomseq_t*)param;
-    long start = SDL_GetTicks64();
-    long delay = 0;
+    long long start = SDL_GetTicks64();
+    long long delay = 0;
     int status;
-    int count = 0;
+    Uint64 count = 0;
     signalhandler signal;
 
     while (1) {
